+++ RELEASE-NOTES.md
@@ -13,6 +13,7 @@
 * #574 Robust ECharts embedding script
 * #576 QL: Can't specify parameters for "col(?)" index
 * #577 Udf0 functional interface
+* #581 Support loading Avro files with "standard" logical types
 
 ## Release 2.0.0-M5